Move seed bank photo code to seedbank package
Since it's all very seed-bank-specific and is unlikely to be too useful as a
starting point for a more general-purpose photo storage system, the code for
reading and writing accession photos should really live in the `seedbank`
package. Once we add a more generic photo storage system, this code can be
removed or turned into a facade for the new code.
