+++ src/main/kotlin/com/terraformation/backend/seedbank/db/PhotoRepository.kt
@@ -1,11 +1,11 @@
-package com.terraformation.backend.photo
+package com.terraformation.backend.seedbank.db
 
 import com.terraformation.backend.config.TerrawareServerConfig
 import com.terraformation.backend.db.AccessionNotFoundException
 import com.terraformation.backend.db.FacilityId
 import com.terraformation.backend.db.tables.daos.AccessionPhotosDao
 import com.terraformation.backend.db.tables.pojos.AccessionPhotosRow
-import com.terraformation.backend.seedbank.db.AccessionStore
+import com.terraformation.backend.seedbank.model.PhotoMetadata
 import java.io.IOException
 import java.io.InputStream
 import java.nio.file.FileAlreadyExistsException
